+++ cordapp-example/contracts-kotlin/src/main/kotlin/com/example/iou/Receita.kt
@@ -6,11 +6,11 @@ import net.corda.core.serialization.CordaSerializable
 
 @CordaSerializable
 data class Receita(
-        val numeroReceita: Int,
+        val numeroReceita: Long,
         val nomePaciente: String,
         val enderecoPaciente: String,
         val nomeMedico: String,
-        val crmMedico: Int,
+        val crmMedico: Long,
         val nomeMedicamento: String,
         val quantidadeMedicamento: Int,
         val formulaMedicamento: String,
